% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{covid19}
\format{An object of class \code{"list"};
\describe{
\item{demo}{20 samples x 4 variables}
\item{mrna}{20 samples x 23,710 transcripts}
}}
\usage{
data(covid19)
}
\description{
COVID-19 datasets: assay characteristics, mRNA transcripts (transcriptomics)
}
\examples{
library(omicsBioAnalytics)
data(covid19)
}
\references{
Blanco-Melo et al. SARS-CoV-2 launches a unique transcriptional signature from in vitro, ex vivo, and in vivo systems. Can J Cardiol. 2019 Apr;35(4):471-479.
(\href{https://www.biorxiv.org/content/10.1101/2020.03.24.004655v1.full}{COVID-19 bioRxiv paper})
}
\keyword{IAV,}
\keyword{RSV}
\keyword{SARS-CoV-2,}
\keyword{bronchial}
\keyword{cells,}
\keyword{epithelial}
\keyword{human}
\keyword{primary}
