% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUploadUI.R
\name{customDownloadButton}
\alias{customDownloadButton}
\title{https://stackoverflow.com/questions/49350509/adding-removing-icon-in-downloadbutton-and-fileinput}
\usage{
customDownloadButton(outputId, label = "Download", icon = icon("download"))
}
\description{
https://stackoverflow.com/questions/49350509/adding-removing-icon-in-downloadbutton-and-fileinput
}
