% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartFailure.R
\docType{data}
\name{heartFailure}
\alias{heartFailure}
\title{heartFailure}
\format{An object of class \code{"list"};
\describe{
\item{demo}{58 subjects x 29 demographic variables}
\item{cells}{58 subjects x 14 cell-types}
\item{holter}{58 subjects x 29 Holter variables}
\item{mrna}{58 subjects x 5000 mRNA transcripts}
\item{proteins}{58 subjects x 65 proteins}
}}
\usage{
data(heartFailure)
}
\description{
Heart failure datasets: demographics, cells, Holter variables, mRNA transcripts, proteins
}
\examples{
library(omicsBioAnalytics)
data(heartFailure)
}
\references{
Singh et al. Ensembling Electrical and Proteogenomics Biomarkers for Improved Prediction of Cardiac-Related 3-Month Hospitalizations: A Pilot Study. Can J Cardiol. 2019 Apr;35(4):471-479.
(\href{https://amritsingh.shinyapps.io/multiomics_HFhospitalizations/}{heart failure datasets})
}
\keyword{biomarkers}
\keyword{blood,}
\keyword{failure,}
\keyword{heart}
\keyword{multi-omics,}
